pyproject.toml
[project]
name = "geniesae"
version = "0.1.0"
description = "Train Sparse Autoencoders on GENIE diffusion language model residual stream activations and evaluate via activation patching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.1.0",
    "nnsight>=0.3.0",
    "datasets>=2.14.0",
    "transformers>=4.35.0",
    "pyyaml>=6.0",
    "numpy>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "hypothesis>=6.90.0",
]

[project.scripts]
geniesae = "geniesae.pipeline:main_cli"
